More matches from group effort on decomp.me

+++ include/main/rng.h
@@ -29,9 +29,9 @@ u32 Rng_Rand32(void);
  * shifts the result right to produce a value within the range
  * of [0, 0x7FFF].
  *
- * @return A random positive 16-bit unsigned integer (u16).
+ * @return A random positive 16-bit unsigned integer (u32).
  */
-u16 Rng_Rand16(void);
+u32 Rng_Rand16(void);
 
 /**
  * @brief Returns the current random seed value.
